Remove Ito World as a provider
Ito World are no longer (and haven't for several years) providing consulting services for organisations interested in OSM data. We still use OSM data internally and for a few services, however don't want to advertise and attract clients for this as it's not a core part of our business. Linked to #53