Add the ".style" to "OSM2PGSQL_OPTIONS"
"src/openstreetmap-carto/openstreetmap-carto.style" has changed, and "addr:interpolation" is no longer stored for nodes.  This makes logical sense and is a bugfix in openstreetmap-carto, but is different to osm2pgsql's default.  This means that "OSM2PGSQL_OPTIONS" needs to reference the .style file that was used to create the database in the first place (arguably, it always should have done this).

The actual location of the .style file depends on the account that the user used to fetch it, and on Debian it'll always be different to the account that renderd runs under.

https://help.openstreetmap.org/questions/78817/keep-in-sync-local-database-with-osm-latest-changes-to-region is an example of this problem elsewhere.