node-carto is back in Debian12 with a suitable version (#242)
See: https://packages.debian.org/bookworm/node-carto

This allows to download a lot less packages from apt as we can skip installing npm, which has a LOT of dependencies. It is also cleaner to install all of the needed tools from Debian, and allows to skip an install step.

I tested it on a fresh Debian 12 install.

    $ apt list node-carto
    Listing... Done
    node-carto/stable,now 1.2.0-4 all [installed]
    $ which carto
    /usr/bin/carto
    $ carto --version
    1.2.0

I compared with the one installed from npm and it emits the same amount of warning.