Serving tiles, Debian 13 This is raster tiles only (all the serving tiles guides are so far).
It is for the "pgsql" output of osm2pgsql, to match what is on osm.org.
A "flex" version (separate, but using much of the same content) would also make sense.

It's a standalone page currently; it'll get added to the menus later.
This is so that people can test it and highlight any errors for correction before it becomes part of the main site.