Serving tiles, Debian 13 (#299)
* Serving tiles, Debian 13
This is raster tiles only (all the serving tiles guides are so far).

It is for the "pgsql" output of osm2pgsql, to match what is on osm.org.
A "flex" version (separate, but using much of the same content) would also make sense.

This PR also adds to the menus.

* Changes following review at
https://github.com/switch2osm/switch2osm/pull/299

* Changed form of words to address comment
https://github.com/switch2osm/switch2osm/pull/299#discussion_r2280526264

* As suggested,
These instructions are have been written and tested against a new {{ dist }} server.

* Spelling ("archtecture") and you need to run updatedb first.

* Grammar: "These instructions have been written"

* Update index.md and mkdocs.yml
Normally I'd make a page "live" before adding to index, but this has had plenty of testing now.

---------

Co-authored-by: Andy Townsend <ajtown@dell5000.atownsend.org.uk>